拆分文件
